requirements.txt
easyocr==1.7.0
pillow==9.5.0
spacy==3.7.2
flask==3.0.3
thefuzz==0.22.1
pypdf==5.0.1
milvus-haystack==0.0.11
haystack-ai==2.6.1
jinja2==3.1.4
markupsafe==2.1.3
itsdangerous==2.2.0
werkzeug==3.0.3
huggingface-hub==0.26.1